QA-7236: moved test runners
